Move implementations - learning layers (#1567)
* move some implementations around
* move base_convolution code to a separate cpp file
