+++ src/layers/learning/channelwise_scale_bias.cpp
@@ -34,9 +34,13 @@ void channelwise_scale_bias_layer<TensorDataType, T_layout, Dev>::fp_compute() {
   using CPUMatType = El::Matrix<TensorDataType, El::Device::CPU>;
 
   // Local matrices
-  const auto& local_input = dynamic_cast<const CPUMatType&>(this->get_local_prev_activations());
-  auto& local_output = dynamic_cast<CPUMatType&>(this->get_local_activations());
-  const auto& local_weights = dynamic_cast<const CPUMatType&>(this->get_data_type_weights(0).get_values().LockedMatrix());
+  const auto& local_input =
+    dynamic_cast<const CPUMatType&>(this->get_local_prev_activations());
+  auto& local_output =
+    dynamic_cast<CPUMatType&>(this->get_local_activations());
+  const auto& local_weights =
+    dynamic_cast<const CPUMatType&>(
+      this->get_data_type_weights(0).get_values().LockedMatrix());
   const auto local_scale = El::LockedView(local_weights,
                                           El::ALL, El::IR(0));
   const auto local_bias = El::LockedView(local_weights,
@@ -69,7 +73,6 @@ void channelwise_scale_bias_layer<TensorDataType, T_layout, Dev>::fp_compute() {
       }
     }
   }
-
 }
 
 template <typename TensorDataType, data_layout T_layout, El::Device Dev>
@@ -77,11 +80,17 @@ void channelwise_scale_bias_layer<TensorDataType, T_layout, Dev>::bp_compute() {
   using CPUMatType = El::Matrix<TensorDataType, El::Device::CPU>;
 
   // Local matrices
-  const auto& local_input = dynamic_cast<const CPUMatType&>(this->get_local_prev_activations());
-  const auto& local_gradient_wrt_output = dynamic_cast<const CPUMatType&>(this->get_local_prev_error_signals());
-  auto& local_gradient_wrt_input = dynamic_cast<CPUMatType&>(this->get_local_error_signals());
-  const auto& local_weights = dynamic_cast<const CPUMatType&>(this->get_data_type_weights(0).get_values().LockedMatrix());
-  auto& local_gradient_wrt_weights = dynamic_cast<CPUMatType&>(this->m_weights_gradient->Matrix());
+  const auto& local_input =
+    dynamic_cast<const CPUMatType&>(this->get_local_prev_activations());
+  const auto& local_gradient_wrt_output =
+    dynamic_cast<const CPUMatType&>(this->get_local_prev_error_signals());
+  auto& local_gradient_wrt_input =
+    dynamic_cast<CPUMatType&>(this->get_local_error_signals());
+  const auto& local_weights =
+    dynamic_cast<const CPUMatType&>(
+      this->get_data_type_weights(0).get_values().LockedMatrix());
+  auto& local_gradient_wrt_weights =
+    dynamic_cast<CPUMatType&>(this->m_weights_gradient->Matrix());
   const auto local_scale = El::LockedView(local_weights,
                                           El::ALL, El::IR(0));
   auto local_gradient_wrt_scale = El::View(local_gradient_wrt_weights,
@@ -128,7 +137,6 @@ void channelwise_scale_bias_layer<TensorDataType, T_layout, Dev>::bp_compute() {
   if (opt != nullptr) {
     opt->add_to_gradient(*this->m_weights_gradient, El::TypeTraits<TensorDataType>::One(), true);
   }
-
 }
 
 #define PROTO(T)                                      \
